+++ pyriemann/utils/__init__.py
@@ -1,0 +1,1 @@
+from .docs import deprecated  # noqa: F401